requirements.txt
#bidict==0.23.1
#blinker==1.7.0
#click==8.1.7
#colorama==0.4.6
Flask==3.0.3
Flask-SocketIO==5.3.6
#future==1.0.0
#h11==0.14.0
#itsdangerous==2.2.0
#Jinja2==3.1.3
#lxml==5.2.1
#MarkupSafe==2.1.5
#pymavlink==2.4.41
pyserial==3.5
pyzmq==26.0.2
requests==2.25.1
#python-engineio==4.9.0
#python-socketio==5.11.2
#pyzmq==19.0.2
#simple-websocket==1.0.0
#Werkzeug==3.0.2
#wsproto==1.2.0
numpy==1.19.5
smbus==1.1.post2
# RPi.GPIO==0.7.0
